chore: add SQLException details to devMessage
The devMessage only contained the SqlState in case there is one. The SqlState is
already part of the message so the devMessage does not contain any more info.

Use the exception message as the devMessage to give more details. When we had
a migration issue leading to a missing table this information was part of
the SQL exception but hidden due to the current logic