fix: fail startup if we cannot create root container
The previous behavior was hiding the root cause by using a fallback that neglects
what the user wanted. This could cause more work on the user side to get data
from the filestore back to the cloud once they notice the issue. Or it might
cause data loss when falling back to transient.