chore: use single program.EnrollmentStatus (#17613)
We had 1:1 mappings in domain and view. The only advantage I see
with separating them is that we would not immediately leak a new status
to the outside world. We do not use such a separation for Event.status.
Event.status is of type org.hisp.dhis.event.EventStatus currently in dhis-api.