test: HibernateCacheBaseTest is not needed as base class
We have too many base test classes that do not provide some common
functionality. They seem to exist because configuring our tests is
complicated and our base test classes and configs are not well
structured and named.

@ContextConfiguration is inherited by default. We can thus inherit the
config of a base class like IntegrationTestBase and override what should
be different. In the case of HibernateQueryCacheTest that is
DhisConfigurationProvider bean as we want to configure the 'hibernate.cache'.

Allowing overriding IntegrationTestBase.before. This allows us to reuse
IntegrationTestBase while changing the BeforeEach setup behavior.

Allow creating a TestDhisConfigurationProvider with properties or
properties from a file. There is no need for this class to be abstract.