fix: bundling apps on CI
the apps where cloned but after the war was packaged so the apps were not in the war.
Run exec in the same phase as we do the frontend plugin. The target folder is not there
at this stage. It was created by the plugin. Now, when running in CI we create it
via the bundle-apps.js script