ci: run all test on mvn test by default
Before we had to add '--activate-profiles -default' to run all tests. That is confusing. The default
is now that we run all types of tests if you do mvn test (unit, integration, integrationH2).
This way we cannot miss any tests in a pipeline due to forgetting to remove the default profile.
On Jenkins we simply do 'mvn clean install' which runs all tests and creates the war. We can do that
since the machines running Jenkins have more resources than the ones on GH. On GitHub we now use
the maven profiles unit-test, integration-test, integration-h2-test to run each of them in a concurrent job.

* AuthTest was not properly classified as an integration test. Now it will run with all and in the integration-test profile.
* Remove unused dev profile.
* Declare maven-surefire-report-plugin and set its version to the.
  maven-surefire-plugin one. We can use it to double check the test
  counts so total of all tests=unit+integration+integrationH2 maven
  profile ones.