ci: don't download webapps when only running tests (#18801)
* ci: don't download webapps when only running tests

Move it the bundling further down in the maven lifecycles
https://maven.apache.org/guides/introduction/introduction-to-the-lifecycle.html\#Lifecycle_Reference
Just before the package phase where the war is built

* ci: do not package war in all test jobs

* test: get Jacoco report when running mvn test

* ci: run dependency check during verify

so it only runs once on GH and not during 'mvn test'. It will
run after the package phase.

Therefore change the run-api-tests to run 'mvn verify'. We run 'mvn
install' on Jenkins.

* test: need to compile tests for maven dep plugin

The maven dependency plugin thinks the deps are unused declared ones if we
run verify without compiling the test classes