test: improve test for exporting tracker event notes
the fix for the issue came in

https://github.com/dhis2/dhis2-core/pull/18943

I tested it in part by asserting the /events/{uid} body is equal to
/events?events={uid}. The test could hide a regression if we were to not
return notes at all in both bodies.

Porting the test from the backport

https://github.com/dhis2/dhis2-core/pull/18949

as it does not suffer from that issue and it points us directly to the
issue.