chore: simplify TrackedEntityService.getTrackedEntity(UID, ...) logic
We have two methods for getting a single TE. They were implemented
independently with convoluted logic. Let one reuse the other. Separate
fetching, ACL, mapping to make the flow easier to follow. Move the TETAV
logic in one place. This is the only way we can ensure the behavior is
the same across different code paths.