chore: allow testing sync using docker-compose (#20297)
the second instance `web-sync` needs to connect to its own DB `db-sync`.
The easiest way is to use a separate `dhis.conf`. Other solutions are
really hacky like only being able to inline the config into the
`docker-compose.yml` if we override the command that starts DHIS2 with a
shell script that adds the config as heredoc.