ci: allow passing .env files as perf test inputs DHIS2-20289 (#22239)
* ci: allow passing .env files as perf test inputs

GitHub workflows have a maximum of 10 inputs which we have already surpassed. This is why I had to
grep for analytics in the JSON fixture name to set an arg for run-simulation.sh.

I recently hit https://github.com/dhis2/dhis2-core/pull/22236 where dhis2/core:42.2 failed to start
with the DB sierra-leone/2.41.0. We need to discuss how to deal with changes that then require user
intervention from implementations. In that case we are the users/maintainers of the SL DBs.

For the time being devs need to be able to set different DBs for baseline and candidate. That means
even more workflow parameters would be needed. A common workaround to the 10 input parameter
limitation is to take in multiline strings and parse them. Lucky for us run-simulation.sh
already spits out a run-simulation.env which you can use locally to reproduce a run. So now
we use the same pattern for the workflow itself.

The workflow should only allow comparing the same Gatling simulation and the same version of it. The
first one is validated in a separate step. The second one ensured by the perf_tests_git_ref input.
So the GIT_ related env vars in a run-simulation.env are ignored.

* docs: readme

* test: fix quotes around args

* ci: warn about quotes